Update checkbox/radio for configurable enclosed option.
The BaseHtml checkbox and radio implementation hard-codes a container to envelop the input when label is passed. For example: ```<div class="checkbox">``` or ```<div class="radio">```. This should be configurable. The current Pull Request allows the container enclosure to be optional by including an ```isEnclosed``` option.This is set to true by default. But it requires to be false for some scenarios. For example, the false option is very useful for generating inline checkbox lists or inline radio lists, where the enclosure is not required.