Implement dropdown menu headers and dividers
Implements headers and dividers in a logical manner similar to Yii 1.1. This allows use of dividers and headers in a more consistent manner (rather than having to manually add HTML), and allows 'visible' and other options to be used.

Items with no URL will become a header. A string with three hyphens will become a divider.
```php
['label' => 'Secret Areas'],
['label' => 'Nimh', 'url' => 'http://www.imdb.com/title/tt0084649'],
'---'
['label' => 'Not Visible', 'visible' => false,
['label' => 'Yii', 'url' => 'http://www.yiiframework.com']
```
will now become:
```html
<li class="dropdown-header">Secret Areas</li>
<li><a tabindex="-1" href="http://www.imdb.com/title/tt0084649">Nimh</a></li>
<li class="divider">
<li><a tabindex="-1" href="http://www.yiiframework.com">Yii</a></li>
```
Original code to output submenus (lines 92, 99-104) retained despite lack of support in Bootstrap 3, as functionality can be restored with a few lines of CSS as shown at http://www.bootply.com/86684 .