Update Menu.php
The feature:
http://getbootstrap.com/components/#list-group-linked

Currently, Menu isnt capable of creating a bootstraps list-group-linked because:
- Menu forces a item container around the link

Menu does allow:
- defining the tag of the menu container
- defining the tag of the item container
- defining the template of the link

This means, that it is a small step to enable a menu like the list-group-linked in which:
- menu container is a DIV
- there are no item containers
- there are only links which should get the correct status by the (link)template

I provided a patch in which you can set the item-tag on false. When this is done,
the item tag is ignored and no LI is created, which is necessary for the list-group-linked.