Added submenuOptions property to Dropdown
A submenu options array adds flexibility to the dropdown, allowing custom attributes to be added to submenus. This patch doesn't break BC, as it sets submenuOptions to options, if it has not been explicitly set. However it does get the rid of the dropdown-menu class that is currently set on submenus, causing them to be hidden by default.