Faster buildInCondition from array of integers
buildInCondition checks each value from an array for it's type(NULL or Expression). Otherwise it converts value to PDO binding param, which is not effective for numbers.
My fix just skips variables of integer or double type so they are not binded, leaved as-is and are just imploded on line 1108. That's much faster.