Initialize input tag value with ng-init when using AngularJS
If AngularJS is used and input tag has mg-model then value will be ignored and input tag will not be initialized with the value.
In order to properly initialize input element we must provide ng-init attribute. 
I propose this enhancement and add ng-init="<model_name>='<value>'", but only if developer requested it 
by setting 'ng-init' option property to true to make it backward compatible.