Update tutorial-core-validators.md
Changing the basic description of the Exist core validator. The previous description lead me to think that an attribute from a Model extending class could not be validated with the Exist validator because it was not an ActiveRecord extending class. 
The truth is that only the exist validator target class has to be extending ActiveRecord. The validated input can be from a class only extending Model.
I hope my changes are sufficiently clear.