Portability fixes for SchemaBuilder.
Removed length from integer and boolean types. This is not supported on PostgreSQL,
and is only used for display in MySQL, not input validation.

Removed length from text and binary type. This is not supported on most RDBMS's and ignored by others.

Changed binary mapping to `varbinary(max)` on MSSQL instead of `binary(1)` (which requires a length and is fixed length)
to better match others.

Changed timestamp mapping to datetime on MSSQL. Timestamp is a unique number on MSSQL, not a timestamp.

Removed precision and scale parameters from floating point types. This is not supported by PostgreSQL, among others.

Renamed timestamp length parameter to precision. The precision parameter is not supported on all RDBMS's, and may
throw an error.