Fix to get data methods and pjax working together
This is a fix for the issue of pjax links/buttons not working when using data methods. This fix also requires a change in the Pjax widget that adds the attribute data-pjax-container to the container element so that it can be found via JavaScript. The pjax container can be defined in two ways, either by setting the data-pjax-container="#someSelector" in the link/button attributes, or else it will find the closest element with data-pjax-container="". So the HTML attribute 'data-pjax-container' is used in two ways: if the value is blank than that means the element is a pjax container, if the value is set then that means it's a selector pointing to a pjax container. If neither are set and the container is not found it will default to the <body> tag.