canonical url needs to ignore specified params
Canonical URL should not capture all the parameters provided in the URL. So we need to specify what parameters to ignore when generating the canonical URL. For example, on a product listing page I would need to ignore a "sort"parameter and point all the sorting links to the one specific product page. 
Ex: 
1. www.mysite.com/mens/dresses?sort=by_price
2. www.mysite.com/mens/dresses?sort=by_date

Canonical URL for above two samples should be.
www.mysite.com/mens/dresses

So I need to tell to ignore 'sort' param when generating the canonical url. That is why this change is proposed.