Update tutorial-yii-integration.md
I found the existing language to be quite confusing. It took me several attempts of several hours each to get it working on my computer. I think my language is a lot clearer and more specific.

However, note that I could not get it to work without commenting out a line of vendor/yiisoft/yii2/base/Object.php, which can't be a good thing. Hopefully someone will know how to avoid this problem and will be able to further correct my instructions.