Fixes false positive on loadFilesRecursive
When running a functional test with codeception, I wasn't getting a null of UploadedFile::getInstance().
Examining further, I realized that the $errors variable type, on elseif, was a string and UPLOAD_ERR_NO_FILE was a integer, therefore the test was returning true ($errors !== UPLOAD_ERR_NO_FILE).

But the $errors value was "4" (string) and UPLOAD_ERR_NO_FILE is 4 (integer).

intval($errors) worked for me and I see no drawbacks on using it.