Guide about Client side Javascript of ActiveForm
I feel that this part is not documented very well and deserves a
dedicated guide page to handle the different possibilities and use
cases.

This covers questions like https://github.com/yiisoft/yii2/issues/11177#issuecomment-200586587

Some parts of what is covered in
https://github.com/samdark/yii2-cookbook/blob/master/book/forms-activeform-js.md
should be covered in the official guide too.

This is WIP, comments and additions welcome.