Allow remove of previous URL when recovered
Customized previous URL is persistent, that makes impossible to condition display of previous link. I explain.

I have a web with products, they are listed in home and in search view. When you click on a product (open product view) in search results i want to show a "back to results" link, but when click on a product in Main page i don´t want to show this link.
The problem is that as the links is stored in session it is always displayed one time it is defined.
This modification allows to clear previous url value just when it is recovered. 
To condition display of backlink in my project i do "if($prev=Url::previous('name')) {//display link}"