Fix number attributes comparison
Assume we have 2 attributes: `min` and `max`. Also we have a rule that `max` is equal or greater than `min` and they both are optional.
Than in case `min` is empty and `max` is set we got a validation error. The reason is that `parseFloat()` function converts empty string of `min` to `NaN`.