incorrect float to string conversion
on my system,  (string) $number in line
   return str_replace(',', '.', (string) $number);
is converted to scientific notation, so it returns this 1.51643E+9 instead of 1516425820.449680. 

By the way this result is being used in \yii\log\Target::getTime(), this is obviously not the expected output of this code. Yet I can not find anything in my php configuration that would explain this. Although, it seems that I am only encountering this when I am using cli. 

I am seeing this on fedora ( latest php) and debian 9, running:

$ php --version
PHP 7.1.12-2+0~20171207160618.12+stretch~1.gbp5c91f3 (cli) (built: Dec  7 2017 16:06:20) ( NTS )
Copyright (c) 1997-2017 The PHP Group
Zend Engine v3.1.0, Copyright (c) 1998-2017 Zend Technologies
    with Zend OPcache v7.1.12-2+0~20171207160618.12+stretch~1.gbp5c91f3, Copyright (c) 1999-2017, by Zend Technologies



it seems number_format() is a more foolproof and setting independent way of accomplishing this conversion