Added dynamic alias '@alias' in ActiveQuery
The @alias placeholder enables to address dynamicaly the alias of the
table name in an active query or in a relation (without knowing it in
advance).

Usage example:
class Order {
...
  public static function find()
  {
    return parent::find()->orderBy( '@alias.id' );
  }

  //relation with table COMPANY
  public function getCustomer()
  {
    return $this->hasOne( Company::className(), ['id' => 'customerid'] )
                ->andOnCondition( [ '@alias.type' => 'customer' ] );
  }
...
}

$orders = Order::find()->alias('o');
//will build the sql:
// select * from order o orderby o.id

$orders = Order::find()->alias('o2')->joinWith('customer c');
//will build the sql:
// select o2.* from order o2 
// left join company c on c.id = o2.customerid and c.type='customer'
// order by o2.id

The dynamic alias feature can be enabled in the application
configuration by the property 'enableAliasDynamic'.
It is false by default so the feature is skipped to avoid possible speed
concerns when not needed.

To enable, you can add the following line in config/main.php:
return [
    'enableAliasDynamic' => true,
];