Update GridView.php
If you do not want to make changes to yii.gridView.js, let at least be able to override GridViewAsset and connect your version of yii.gridView.js there. Because in our application there are no default filters that need to be rewritten when filtering, but the problem with long URLs after parameter substitution when using filtering is very relevant for us. Thank you in advance!