Rectifying wrong instruction
Used to say:
running the following console command while in the project `web` directory

Now says:
running the following
console command while in the project root directory

Why:
I tried it, the `web` directory didn't work.

I'm brand new to yii, so to double check if it was an issue with my environment I asked an expert on it. He stated that the yii command-line tool is in the root directory and that it isn't possible to have a new basic install run in the `web` directory if you `curl`ed it just before. 

If it is possible to run it with the `web` directory with other installation methods, then both folders should be suggested.