feat(cache): Stopped generating constant directory level
Removed key prefix for result of `\yii\caching\FileCache::getCacheFile()` to avoid generating constant cache directory levels

BREAKING CHANGE: After the Yii upgrade, the storage structure for the
file cache will change. It is worth warming up the cache again if there
is a logical dependency when working with the file cache.

Closes https://github.com/yiisoft/yii2/issues/18263