Handles situations with more complex ICU unit data structure
It turned out that ICU unit data in the new version 7.1+ (at least I found this in PHP 8.2, where the libraries are version 7.1) have a more complex structure for some languages (in my case cs_CZ). In this case, the code tries to go through and use a structure that is not ready for it.